+++ src/editor/components/filter/QueryEditorFilterSection.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { QueryEditorOperatorDefinition, QueryEditorCondition, QueryEditorFieldDefinition } from '../../types';
 import { QueryEditorSectionRenderer } from '../QueryEditorSectionRenderer';
 import { QueryEditorSectionBaseProps, QueryEditorSection } from '../QueryEditorSection';
-import { QueryEditorExpression, QueryEditorSectionExpression } from '../../../types';
 import { SelectableValue } from '@grafana/data';
 import { ExpressionSuggestor } from '../types';
+import { QueryEditorExpression, QueryEditorSectionExpression } from '../../expressions';
 
 interface FilterSectionConfiguration {
   id: string;
