+++ src/editor/components/reduce/QueryEditorReduceSection.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { QueryEditorFieldDefinition } from '../../types';
 import { QueryEditorSection, QueryEditorSectionBaseProps } from '../QueryEditorSection';
-import { QueryEditorExpression, QueryEditorSectionExpression } from '../../../types';
 import { QueryEditorSectionRenderer } from '../QueryEditorSectionRenderer';
 import { SelectableValue } from '@grafana/data';
 import { ExpressionSuggestor } from '../types';
+import { QueryEditorExpression, QueryEditorSectionExpression } from '../../expressions';
 
 interface ReduceSectionConfiguration {
   id: string;
