+++ src/editor/components/types.ts
@@ -1,4 +1,4 @@
-import { QueryEditorExpression } from '../../types';
 import { SelectableValue } from '@grafana/data';
+import { QueryEditorExpression } from '../expressions';
 
 export type ExpressionSuggestor = (txt: string, skip: QueryEditorExpression) => Promise<Array<SelectableValue<string>>>;
