+++ src/editor/expressions.ts
@@ -1,5 +1,24 @@
-import { QueryEditorExpression, QueryEditorExpressionType } from '../types';
 import { QueryEditorFieldType, QueryEditorOperatorDefinition } from './types';
+
+export enum QueryEditorExpressionType {
+  Field = 'field',
+  Conditional = 'conditional',
+  Operator = 'operator',
+  FieldAndOperator = 'fieldAndOperator',
+  OperatorRepeater = 'operatorRepeater',
+  Reduce = 'reduce',
+  FunctionParameter = 'functionParameter',
+  GroupBy = 'groupBy',
+}
+
+export interface QueryEditorExpression {
+  type: QueryEditorExpressionType;
+}
+
+export interface QueryEditorSectionExpression {
+  id: string;
+  expression?: QueryEditorExpression;
+}
 
 /** OPERATORS */
 export interface QueryEditorOperatorExpression extends QueryEditorExpression {
